refactor: modify navigation actions for categories
