+++ app/src/main/java/com/moscow/tudee/presentation/category/CategoriesEvents.kt
@@ -1,0 +1,5 @@
+package com.moscow.tudee.presentation.category
+
+sealed class CategoriesEvents {
+    data class NavigateToTasks(val categoryId:Long) : CategoriesEvents()
+}
